Change date display in project view
When the due date is yesterday, today, or tomorrow, it will be written
as such instead of writing the exact date. Additionally, the date color
will change in red if the date comes before today to indicate that the
deadline has been passed.There is currently a bug with date-fns which
does not retrieve the date properly around midnight, but this is out of
scope for this project.